refactor: remove the PortSelectionComponent to the form directory
